Fix linking errors on the JS platform.
The code that parallelizes the execution of tests is extracted in
a platform-dependent Platform object. On the JS, it systematically
uses only one worker, which avoids to refer to ForkJoinPool and
other APIs of parallelism that are not supported on the JS platform.
