+++ js/src/main/scala/org/scalacheck/Platform.scala
@@ -1,0 +1,24 @@
+/*-------------------------------------------------------------------------*\
+**  ScalaCheck                                                             **
+**  Copyright (c) 2007-2015 Rickard Nilsson. All rights reserved.          **
+**  http://www.scalacheck.org                                              **
+**                                                                         **
+**  This software is released under the terms of the Revised BSD License.  **
+**  There is NO WARRANTY. See the file LICENSE for the full text.          **
+\*------------------------------------------------------------------------ */
+
+package org.scalacheck
+
+import Test._
+
+private[scalacheck] object Platform {
+
+  def runWorkers(
+    params: Parameters,
+    workerFun: Int => Result,
+    stop: () => Unit
+  ): Result = {
+    workerFun(0)
+  }
+
+}
