+++ src/main/scala/org/scalacheck/Test.scala
@@ -345,45 +345,8 @@ object Test {
       } else res
     }
 
-    def mergeResults(r1: Result, r2: Result): Result = {
-      val Result(st1, s1, d1, fm1, _) = r1
-      val Result(st2, s2, d2, fm2, _) = r2
-      if (st1 != Passed && st1 != Exhausted)
-        Result(st1, s1+s2, d1+d2, fm1++fm2, 0)
-      else if (st2 != Passed && st2 != Exhausted)
-        Result(st2, s1+s2, d1+d2, fm1++fm2, 0)
-      else {
-        if (s1+s2 >= minSuccessfulTests && maxDiscardRatio*(s1+s2) >= (d1+d2))
-          Result(Passed, s1+s2, d1+d2, fm1++fm2, 0)
-        else
-          Result(Exhausted, s1+s2, d1+d2, fm1++fm2, 0)
-      }
-    }
-
     val start = System.currentTimeMillis
-
-    val r =
-      if(workers < 2) workerFun(0)
-      else {
-        import concurrent._
-        val tp = java.util.concurrent.Executors.newFixedThreadPool(workers)
-        implicit val ec = ExecutionContext.fromExecutor(tp)
-        try {
-          val fs = List.range(0,workers) map (idx => Future {
-            params.customClassLoader.map(
-              Thread.currentThread.setContextClassLoader(_)
-            )
-            blocking { workerFun(idx) }
-          })
-          val zeroRes = Result(Passed,0,0,FreqMap.empty[Set[Any]],0)
-          val res = Future.fold(fs)(zeroRes)(mergeResults)
-          Await.result(res, concurrent.duration.Duration.Inf)
-        } finally {
-          stop = true
-          tp.shutdown()
-        }
-      }
-
+    val r = Platform.runWorkers(params, workerFun, () => stop = true)
     val timedRes = r.copy(time = System.currentTimeMillis-start)
     params.testCallback.onTestResult("", timedRes)
     timedRes
